Operator can configure the behavior of the audit, message queue, password validator, and password authenticator creation with respect to whether singleton instances of these subsystems should be used.
Operator can configure if the password authenticator subsystem should be instantiated as a singleton. Fixes #19
Operator can configure if the password validator subsystem should be instantiated as a singleton. Fixes #18
Operator can configure if the message queue subsystem should be instantiated as a singleton. Fixes #17
Operator can configure if the audit subsystem should be instantiated as a singleton. Fixes #16